requirements.txt
flask
scikit-learn
pandas
google-generativeai